+++ src/components/Avatar/AvatarProps.ts
