+++ src/components/PostCard/PostCardTypes.ts
